[bphh-1798] WIP collaborators sidebar. And implement designated submitter selector
