+++ app/frontend/components/domains/permit-application/collaborator-management/block-collaborator-assignment-management.tsx
