+++ app/frontend/components/domains/permit-application/collaborator-management/types.ts
@@ -1,0 +1,5 @@
+export enum EAssignmentPopoverScreen {
+  collaborations = "collaborations",
+  collaborationAssignment = "collaborationAssignment",
+  collaboratorInvite = "collaboratorInvite",
+}
